py_gcs_bq/requirements.txt
google-cloud-storage
google-cloud-bigquery
python-dotenv
requests
